Refactor: Extract platform detection logic into separate utility functions
